TAO-6804 Moved integration tests to integration test folder from unit test
